Add complexity table, some work with exercises add and showing

+++ app/Http/Livewire/ExerciseCreateComponent.php
@@ -4,31 +4,36 @@ namespace App\Http\Livewire;
 
 use Livewire\Component;
 use App\Models\Exercise;
+use App\Models\Complexity;
 use App\Models\Link;
 
 class ExerciseCreateComponent extends Component
 {
     public $block;
     public $exercise;
-    public $stepFrame = 1;
+    public $complexities;
+    public $levels = [];
+    public $complexity_id = 1;
     public $exercise_title;
     public $exercise_excerpt;
     public $exercise_body;
-    public $validatedData;
+    public $exercise_time;
+
+    public $stepFrame = 1;
 
     public $links = [];
     public $link_name;
     public $link_url;
 
-    public $videos = [];
     public $files = [];
 
     public function toStageTwo()
     {
-        $this->validatedData = $this->validate([
+        $this->validate([
             'exercise_title' => 'required',
             'exercise_excerpt' => 'required',
             'exercise_body' => 'required',
+            'exercise_time' => 'required|numeric|min:10',
         ], [
             'exercise_title.required' => 'Заголовок обязателен для задания',
             'exercise_excerpt.required' => 'Краткое описание обязательно',
@@ -39,6 +44,8 @@ class ExerciseCreateComponent extends Component
             'title' => $this->exercise_title,
             'excerpt' => $this->exercise_excerpt,
             'body' => $this->exercise_body ?: '',
+            'complexity_id' => $this->complexity_id,
+            'time' => $this->exercise_time,
             'user_id' => auth()->user()->id,
             'block_id' => $this->block->id,
         ]);
@@ -60,6 +67,7 @@ class ExerciseCreateComponent extends Component
             'link_url.required' => 'Укажите ссылку чтобы по ней можно было прейти',
             'link_url.url' => 'Поле URL должно быть ссылкой',
         ]);
+
         Link::create([
             'name' => $this->link_name,
             'url' => $this->link_url,
@@ -73,16 +81,25 @@ class ExerciseCreateComponent extends Component
         $this->link_url = '';
 
         $this->links = Link::where('exercise_id', $this->exercise->id)->get();
-        // foreach($this->links as $link) {
-        //     dd($link);
-        // }
+
         session()->flash('message', 'Ссылка к упражнению создана, создадим ещё?');
 
     }
-    public function deleteLink(Link $link)
+
+    public function deleteLink($link)
     {
-        $link->delete();
-        $this->links = $this->exercise->links;
+        $link = Link::find($link);
+        if(!isset($link)) {
+            $this->links = Link::where('exercise_id', $this->exercise->id)->get();
+            session()->flash('error', 'Что то не так... :(');
+        } else {
+            $link->delete();
+            $this->links = Link::where('exercise_id', $this->exercise->id)->get();
+        }
+
+        $this->links = Link::where('exercise_id', $this->exercise->id)->get();
+
+        session()->flash('info', 'Ссылка была успешно удалена :) Не хотите добавить другую?');
     }
 
     // TODO: make edit modal window.
@@ -114,10 +131,7 @@ class ExerciseCreateComponent extends Component
 
     public function completeExerciseCreate()
     {
-        $this->exercise->update([
-            'exercise_body' => $this->exercise_body,
-        ]);
-        $this->stepFrame++;
+        return redirect()->route('admin.tracks.blocks.show',[ $this->block->track_id, $this->block->id]);
     }
 
     // public function mount()
@@ -126,6 +140,11 @@ class ExerciseCreateComponent extends Component
     // }
     public function render()
     {
+
+        $this->complexities = Complexity::all();
+        foreach ($this->complexities as $complexity) {
+            $this->levels[$complexity->id] =  $complexity->level;
+        }
         return view('livewire.admin.exercises.exercise-create-component');
     }
 }
